New Transform
Hierarchical Transforms working on Lazy Update, which means as long as not needed, the global Transform won't be updated. Therefore, children's change of transform won't cause ancestor's, and ancestor's update won't influence children's immediately, but only when one of it's children query it's global transform, the transforms between the changed ancestor and the child will be recalculated.

Briefs:

1. As long as no querying, a global transform won't be recalculated.

2. As long as there's no changes(see 4), a global transform won't be recalculated.

3. Updating a global transform uses it's parent's global transform and the local transform of itself, which causes a query to the parent's global transform.

4. There is changes means one of it's ancestor or itself have changed since the last time it was updated.