Material and Normal mapping
Added a material class which stores a property buffer and two textures(diffuse map and normal map). This new class took the place of the previous Texture class which can only and must store one single texture. With the help of the property buffer, the material can have a diffuse texture, a normal texture, both of them or none of them.

Implemented normal mapping, and calculating of TBN vectors. By disabling a material's diffuse map, you can see the details of shape added by the normal map.