std module
The project now is consuming std module by `import std;` and a bunch of macros to disable `#include <whatever>`. The compile time is now significantly improved.

The bug fix in the last commit was cancelled with the consideration of performance. Another way to prevent deleting a model which is being drawn should be find in the future.