Mesh Builder
Implemented a helper class `MeshBuilder` which stores the editable vertices and triangles and can generate  `MeshModel`s which currently can't be modified once be created and transfered to GPU memory. Also implemented two static functions of `MeshBuilder` to generate boxes and uv spheres which are most often used. More preset models are planned.