Attach Transformation
Attach Transformation is the grabbed object's transform in controller space. This transform is calculated when grabbed if `freeGrabbing = true`. So each frame the grabber will set the grabbed object's global transform as `controllerTransform * attachTransformation`. Of course a fixed `attachTransformation` can be set if `freeGrabbing = false`. Also by inheriting from `XrGrabable`, a more customized setting procedure of
 `attachTransformation` can be implemented.