Grabable Action
A simple grabable action. In `XrGrabable` class provided three `std::function`s to act as grabable action. So there's no need to inherit from `XrGrabable` class and override the virtual functions. Just passing some lambda expressions is enough.

As a example, now player can pick up the revolver and pull trigger to shoot. The revolver is a example for both fixed attachTransformation which is set to the revolver's handle, and grabable action as explained above.

By the way, we come from, where the revolver is bind to the hand position and the bullets are generated by main Loop, to where the revolver became a pickable item and can generate bullets by itself!