Collision filter
Created custom `SimulationFilterShader` based on rules using `PxShape`'s `SimulationFilterData`:
word0: mask.
word1: bits to filter out if counter's mask contains any.
word2: bits that counter's mask must contains all.
word3: friend mask, ignore word0, word1 and word2 as long as two friend masks overlap.

As a example, the game has a hovering white ball called `Self Controller`, with the help of grab actions, player should grab this ball to move in the world instead of always be able to do so. The ball will come back to where it was after released, and won't collide with anything as it's not supposed to.