Text rendering
A rough but working, text rendering pipeline. In fact the shaders of this pipeline is slightly more complex that normal mesh rendering pipeline, because this implementation also takes material and light into account, and have additional `bitmap` need to process. So maybe a more simple implementation which only take's color and bitmap as input, should be planned.